Add Post edit & delete

+++ app/Http/Controllers/PostController.php
@@ -11,21 +11,30 @@ class PostController extends Controller
     {
         $comments = $post->comments;
         $post->increment('views_count');
-        return view('post.show', compact('post','comments'));
+        return view('post.show', compact('post', 'comments'));
+    }
+
+    public function edit(Post $post)
+    {
+        $this->authorize('update', $post);
+        return view('post.edit', compact('post'));
     }
 
     public function update(Request $request, Post $post)
     {
-        $request->validate([
-            'title' => 'nullable|min:5|max:255',
-            'body' => 'required|min:5|max:255',
-            'image' => 'nullable|image|max:1024'
+        $this->authorize('update', $post);
+        $attributes = $request->validate([
+            'body' => 'required|min:5',
         ]);
+
         $post->update($attributes);
+        return redirect('/post/' . $post->id);
     }
 
-    public function delete(Request $request, Post $post)
+    public function destroy(Request $request, Post $post)
     {
+        $this->authorize('delete', $post);
         $post->delete();
+        return redirect('/user/hero/' . $post->postable->id);
     }
 }
